converted the help center page in server comp
